+++ lib/fbutton.dart
@@ -1,3 +1,5 @@
+import 'package:fcontrol/fcontrol.dart';
+import 'package:fcontrol/fdefine.dart';
 import 'package:flutter/material.dart';
 
 /// [FButton] 图片相对与文字的位置
@@ -8,30 +10,21 @@ enum ImageAlignment {
   ///
   /// Left of text
   left,
+
   /// 文字顶部
   ///
   /// Top of text
   top,
+
   /// 文字右边
   ///
   /// Right side of text
   right,
+
   /// 文字底部
   ///
   /// Bottom of text
   bottom,
-}
-
-/// 圆角风格。
-/// [round] - 圆角
-/// [bevel] - 斜角
-///
-/// Rounded corner style.
-/// [round]-rounded corners
-/// [bevel]-beveled corners
-enum FButtonCornerStyle {
-  round,
-  bevel,
 }
 
 /// FButton 提供了一个常用按钮效果的实现集合。支持配置圆角、各种特效、边角、边框以及 Loading等。
@@ -41,247 +34,259 @@ enum FButtonCornerStyle {
 /// Support for configuring rounded corners, various special effects, corners, borders, and loading.
 /// FButton allows developers to familiarize themselves with just one button component and implement common buttons.
 /// No need to master many different types of button components.
+// ignore: must_be_immutable
 class FButton extends StatefulWidget {
   /// 当按下按钮时，会触发该函数。如果未设置该函数，按钮将进入不可用状态。
   ///
   /// When the button is pressed, the function is triggered.If this function is not set, the button will enter an unavailable state.
-  VoidCallback onPressed;
+  final VoidCallback onPressed;
 
   /// 按钮上的文字。
   ///
   /// The text on the button.
-  String text;
+  final String text;
 
-  /// 按钮文字的颜色。
-  ///
-  /// The color of the button text.
-  Color textColor;
-
-  /// 按钮文字的大小
-  ///
-  /// Button text size
-  double fontSize;
-
-  /// 按钮文字的风格
+  /// 按钮文本样式
   ///
   /// Button text style
-  FontStyle fontStyle;
+  final TextStyle style;
 
-  /// 一行文字高度 = textSize * fontHeight
+  /// 禁用按钮文本样式
   ///
-  /// height of one line of text = textSize * fontHeight
-  double fontHeight;
+  /// Disable button text style
+  final TextStyle disableStyle;
 
-  /// 文字粗细。默认 [FontWeight.normal]
+  /// 文本在组件中的相对位置。[Alignment]
   ///
-  /// Text weight. Default [FontWeight.normal]
-  FontWeight textWeight;
-
-  /// 按钮文字的不可用颜色
-  ///
-  /// Unavailable colors for button text
-  Color disabledTextColor;
+  /// The relative position of the text in the component. [Alignment]
+  final Alignment alignment;
 
   /// 按钮的颜色
   ///
   /// Button color
-  Color color;
+  final Color color;
 
   /// 按钮不可用颜色
   ///
   /// Button is not available in color
-  Color disabledColor;
+  final Color disabledColor;
 
-  /// 当 [effect] 为 true 时，触摸按钮时涟漪的颜色。
+  /// 鼠标进入按钮范围时，按钮的颜色。
   ///
-  /// When [effect] is true, the color of the ripple when the button is touched.
-  Color splashColor;
+  /// the color of the button when the mouse enters the button range.
+  final Color hoverColor;
 
-  /// 当 [effect] 为 true 时，鼠标进入按钮范围时，按钮的颜色。
+  /// 按压按钮时的按钮颜色。
   ///
-  /// When [effect] is true, the color of the button when the mouse enters the button range.
-  Color hoverColor;
+  /// the button color when the button is pressed.
+  final Color highlightColor;
 
-  /// 当 [effect] 为 true 时，按压按钮时的按钮颜色。
+  /// 按压按钮时的蒙层颜色。调整颜色值的 alpha，以确保背后的视图能够展示。
   ///
-  /// When [effect] is true, the button color when the button is pressed.
-  Color highlightColor;
+  /// The color of the mask when the button is pressed. Adjust the alpha of the color value to ensure that the view behind can be displayed.
+  final Color activeMaskColor;
 
   /// 内间距。
   ///
   /// Internal spacing
-  EdgeInsetsGeometry padding;
-
-  /// 焦点。
-  ///
-  /// focus
-  FocusNode focusNode;
-
-  /// 是否能够自动获取焦点
-  ///
-  /// Whether it can automatically obtain focus
-  bool autofocus;
+  final EdgeInsetsGeometry padding;
 
   /// 宽度。
   ///
   /// width
-  double width;
+  final double width;
 
   /// 高度。
   ///
   /// height.
-  double height;
+  final double height;
 
-  /// 设置组件圆角。详见 [FButtonCorner]
+  /// 为组件设置边角。
   ///
-  /// Sets the component fillet. See [FButtonCorner] for details
-  FButtonCorner corner;
+  /// Set corners for widget
+  final FCorner corner;
 
-  /// 设置圆角风格，默认 [FButtonCorner.round]
+  /// 设置边角风格，默认 [FCornerStyle.round]
   ///
-  /// Set rounded corner style, default [FButtonCorner.round]
-  FButtonCornerStyle cornerStyle;
+  /// Set rounded corner style, default [FCornerStyle.round]
+  final FCornerStyle cornerStyle;
 
   /// 设置边框颜色。
   ///
   /// Set the border color.
-  Color strokeColor;
+  final Color strokeColor;
 
   /// 设置边框宽
   ///
   /// Set border width
-  double strokeWidth;
+  final double strokeWidth;
 
   /// 设置组件阴影颜色
   ///
   /// Set component shadow color
-  Color shadowColor;
+  final Color shadowColor;
+
+  /// 开启 Neumorphism 风格后的，亮部阴影颜色
+  ///
+  /// After the Neumorphism style is turned on, the bright shadow color
+  final Color highlightShadowColor;
 
   /// 设置组件阴影偏移
   ///
   /// Set component shadow offset
-  Offset shadowOffset;
+  final Offset shadowOffset;
 
   /// 设置组件高斯与阴影形状卷积的标准偏差。
   ///
   /// Sets the standard deviation of the component's Gaussian convolution with the shadow shape.
-  double shadowBlur;
-
-  /// 是否开启特效。开启特效后，[splashColor]、[highlightColor]、[hoverColor] 等属性才有效。
-  ///
-  /// Whether to enable special effects. After the special effects are turned on, the attributes such as [splashColor], [highlightColor], [hoverColor] are effective.
-  bool effect;
+  final double shadowBlur;
 
   /// 是否开启基于阴影的点击特效。
   ///
   /// Whether to enable shadow-based click effects.
-  bool clickEffect;
+  final bool clickEffect;
 
   /// 设置组件渐变色背景。会覆盖 [color] 配置
   /// 你可选择 [LinearGradient]，[RadialGradient]，[SweepGradient] 等..
   ///
   /// Sets the gradient background of the component. [BackgroundColor]
   /// You can choose [LinearGradient], [RadialGradient], [SweepGradient], etc ..
-  Gradient gradient;
+  final Gradient gradient;
 
   /// 设置图标
   ///
   /// Settings icon
-  Widget image;
+  final Widget image;
 
   /// 设置图标与文本的间距
   ///
   /// Set the distance between the icon and the text
-  double imageMargin;
+  final double imageMargin;
 
   /// 设置图标与文本的相对位置。详见 [ImageAlignment]
   ///
   /// Set the relative position of the icon and the text. See [ImageAlignment] for details
-  ImageAlignment imageAlignment;
+  final ImageAlignment imageAlignment;
 
   /// 是否启动 Loading 状态。Loading 状态会覆盖 [image] 配置
   ///
   /// Whether to start the loading state. Loading status will override [image] configuration
-  bool loading;
+  final bool loading;
+
+  /// loading 状态时的 Loading 小部件。会覆盖默认的 Loading 效果
+  ///
+  /// Loading widget in loading state. Will override the default Loading effect
+  final Widget loadingWidget;
 
   /// Loading 的颜色
   ///
   /// Loading colors
-  Color loadingColor;
+  final Color loadingColor;
 
   /// Loading 的宽度
   ///
   /// Loading width
-  double loadingStrokeWidth;
+  final double loadingStrokeWidth;
 
   /// Loading 的大小
   ///
   /// Loading size
-  double loadingSize;
+  final double loadingSize;
 
   /// 是否启用点击进入 Loading 状态的模式
   ///
   /// Whether to enable click to enter the loading mode
-  bool clickLoading;
+  final bool clickLoading;
 
   /// Loading 状态下是否隐藏文本
   ///
   /// Whether to hide text in the loading state
-  bool hideTextOnLoading;
+  final bool hideTextOnLoading;
 
   /// Loading 状态下展示的文本
   ///
   /// Text displayed under Loading
-  String loadingText;
+  final String loadingText;
 
-  double cacheShadowBlur;
+  /// 表面的风格。默认 [FSurface.Flat]。详见 [FSurface]
+  ///
+  /// Surface style. Default [FSurface.Flat]. See [FSurface] for details
+  final FSurface surfaceStyle;
 
-  Color cacheShadowColor;
+  /// 鼠标进入/退出组件范围时会回调
+  ///
+  /// Callback when the mouse enters/exits the component range
+  final ValueChanged<bool> onHover;
 
-  bool get enabled => onPressed != null && !loading;
+  /// 按下时会回调
+  ///
+  /// Callback when pressed
+  final VoidCallback onPressedDown;
 
-  FButton(
-      {Key key,
-      @required this.onPressed,
-      this.text,
-      this.textColor,
-      this.disabledTextColor,
-      this.fontSize,
-      this.fontStyle,
-      this.color,
-      this.disabledColor,
-      this.hoverColor,
-      this.highlightColor,
-      this.splashColor,
-      this.effect = false,
-      this.padding,
-      this.focusNode,
-      this.autofocus = false,
-      this.width,
-      this.height,
-      this.corner,
-      this.cornerStyle = FButtonCornerStyle.round,
-      this.strokeColor,
-      this.strokeWidth,
-      this.shadowColor,
-      this.shadowOffset,
-      this.shadowBlur = 1.0,
-      this.gradient,
-      this.image,
-      this.imageMargin = 6.0,
-      this.imageAlignment = ImageAlignment.left,
-      this.loading = false,
-      this.loadingColor,
-      this.loadingStrokeWidth = 4.0,
-      this.clickLoading = false,
-      this.hideTextOnLoading = false,
-      this.loadingText,
-      this.loadingSize = 12,
-      this.clickEffect = false})
-      : cacheShadowBlur = shadowBlur,
-        cacheShadowColor = shadowColor,
-        assert(autofocus != null),
-        super(key: key);
+  /// 抬起时会回调
+  ///
+  /// Callback when lifted
+  final VoidCallback onPressedUp;
+
+  /// 按下取消时会回调
+  ///
+  /// Callback when cancel is pressed
+  final VoidCallback onPressedCancel;
+
+  /// 是否支持 Neumorphism 风格。开启该项 [highlightColor] 将会失效
+  ///
+  /// Whether to support the Neumorphism style. Open this item [highlightColor] will be invalid
+  final bool isSupportNeumorphism;
+
+  /// 当 [isSupportNeumorphism] 为 true 时有效。光源方向，分为左上、左下、右上、右下四个方向。用来控制光源照射方向，会影响高亮方向和阴影方向
+  ///
+  /// Valid when [isSupportNeumorphism] is true. The direction of the light source is divided into four directions: upper left, lower left, upper right, and lower right. Used to control the illumination direction of the light source, which will affect the highlight direction and shadow direction
+  final FLightOrientation lightOrientation;
+
+  FButton({
+    Key key,
+    this.onPressed,
+    this.text,
+    this.color = Colors.transparent,
+    this.disabledColor,
+    this.hoverColor,
+    this.highlightColor,
+    this.padding,
+    this.width,
+    this.height,
+    this.corner,
+    this.cornerStyle = FCornerStyle.round,
+    this.strokeColor,
+    this.strokeWidth,
+    this.shadowColor,
+    this.shadowOffset,
+    this.shadowBlur = 0.0,
+    this.gradient,
+    this.image,
+    this.imageMargin = 6.0,
+    this.imageAlignment = ImageAlignment.left,
+    this.loading = false,
+    this.loadingColor,
+    this.loadingStrokeWidth = 4.0,
+    this.clickLoading = false,
+    this.hideTextOnLoading = false,
+    this.loadingText,
+    this.loadingSize = 12,
+    this.clickEffect = false,
+    this.style,
+    this.disableStyle,
+    this.alignment,
+    this.activeMaskColor = Colors.transparent,
+    this.surfaceStyle = FSurface.Flat,
+    this.onHover,
+    this.onPressedDown,
+    this.onPressedUp,
+    this.onPressedCancel,
+    this.isSupportNeumorphism = false,
+    this.highlightShadowColor,
+    this.loadingWidget,
+    this.lightOrientation = FLightOrientation.LeftTop,
+  }) : super(key: key);
 
   @override
   State<StatefulWidget> createState() {
@@ -289,92 +294,72 @@ class FButton extends StatefulWidget {
   }
 }
 
-class _FButton extends State<FButton> {
-  final Set<MaterialState> _states = <MaterialState>{};
+class _FButton extends State<FButton> with SingleTickerProviderStateMixin {
+  double shadowBlur;
 
-  bool get _hovered => _states.contains(MaterialState.hovered);
+  Color shadowColor;
 
-  bool get _pressed => _states.contains(MaterialState.pressed);
+  Tween shadowTween;
 
-  bool get _disabled => _states.contains(MaterialState.disabled);
+  AnimationController animationController;
 
-  void _updateState(MaterialState state, bool value) {
-    value ? _states.add(state) : _states.remove(state);
-  }
+  bool get enabled => widget.onPressed != null && !loading;
+
+  bool loading = false;
 
   @override
   void initState() {
     super.initState();
-    _updateState(MaterialState.disabled, !widget.enabled);
+    updateParam();
+    animationController = AnimationController(
+      duration: Duration(milliseconds: 300),
+      vsync: this,
+    );
+    shadowTween = Tween(begin: 0, end: widget.shadowBlur);
+    animationController.addListener(() {
+      setState(() {
+        shadowBlur = shadowTween.evaluate(animationController);
+      });
+    });
   }
 
   @override
   void didUpdateWidget(FButton oldWidget) {
     super.didUpdateWidget(oldWidget);
-    _updateState(MaterialState.disabled, !widget.enabled);
-    if (_disabled && _pressed) {
-      _handleHighlightChanged(false);
-    }
+    updateParam();
   }
 
-  void _handleHighlightChanged(bool value) {
-    if (_pressed != value) {
-      setState(() {
-        _updateState(MaterialState.pressed, value);
-      });
-    }
+  @override
+  void dispose() {
+    animationController.dispose();
+    super.dispose();
   }
 
-  void _handleHoveredChanged(bool value) {
-    if (_hovered != value) {
-      setState(() {
-        _updateState(MaterialState.hovered, value);
-      });
-    }
+  updateParam() {
+    shadowBlur = widget.shadowBlur;
+    shadowColor = widget.shadowColor;
+    loading = widget.loading;
   }
 
   @override
   Widget build(BuildContext context) {
-    var themeData = Theme.of(context);
-    final ButtonThemeData buttonTheme = themeData.buttonTheme;
-    final TextStyle textStyle = themeData.textTheme.button;
-    bool gradientEnable = widget.enabled && widget.gradient != null;
-    double disableOpacity = 0.68;
-
-    /// determine whether to enable special effects
-    widget.effect = widget.enabled ? widget.effect : false;
-    widget.color = widget.enabled
-        ? widget.color ?? buttonTheme.colorScheme.background
-        : widget.disabledColor ??
-            (widget.color == null
-                ? buttonTheme.colorScheme.onSurface.withOpacity(disableOpacity)
-                : widget.color.withOpacity(disableOpacity));
-    Color _textColor = widget.enabled
-        ? widget.textColor ?? textStyle.color
-        : widget.disabledTextColor ??
-            (widget.textColor == null
-                ? buttonTheme.colorScheme.onSurface.withOpacity(disableOpacity)
-                : widget.textColor.withOpacity(disableOpacity));
+    double disableOpacity = 0.58;
 
     /// Handle the relationship between [image] and [loading]
-    widget.image = widget.loading
-        ? SizedBox(
-            width: widget.loadingSize,
-            height: widget.loadingSize,
-            child: CircularProgressIndicator(
-              strokeWidth: widget.loadingStrokeWidth,
-              valueColor: AlwaysStoppedAnimation<Color>(widget.loadingColor ==
-                      null
-                  ? Theme.of(context).accentColor.withOpacity(disableOpacity)
-                  : widget.loadingColor.withOpacity(disableOpacity)),
-            ),
-          )
+    Widget image = loading
+        ? widget.loadingWidget ??
+            SizedBox(
+              width: widget.loadingSize,
+              height: widget.loadingSize,
+              child: CircularProgressIndicator(
+                strokeWidth: widget.loadingStrokeWidth,
+                valueColor: AlwaysStoppedAnimation<Color>(widget.loadingColor ==
+                        null
+                    ? Theme.of(context).accentColor.withOpacity(disableOpacity)
+                    : widget.loadingColor.withOpacity(disableOpacity)),
+              ),
+            )
         : widget.image;
-
-    /// Handle the relationship between loading text and regular text
-    widget.loadingText =
-        widget.loadingText == null ? widget.text : widget.loadingText;
-    widget.text = widget.loading ? widget.loadingText : widget.text;
 
     /// Corner
     BorderRadius borderRadius = widget.corner == null
@@ -387,104 +372,148 @@ class _FButton extends State<FButton> {
           );
 
     /// side
-    var sideColor = widget.enabled
-        ? widget.strokeColor ?? Colors.transparent
-        : widget.strokeColor == null
-            ? Colors.transparent
-            : widget.strokeColor.withOpacity(disableOpacity);
-    var borderSide = BorderSide(
+    Color sideColor = widget.strokeColor ?? Colors.transparent;
+    BorderSide borderSide = BorderSide(
       width: widget.strokeWidth ?? 0,
       color: sideColor,
       style: BorderStyle.solid,
     );
 
     /// shape
-    var shape = widget.cornerStyle == FButtonCornerStyle.round
-        ? RoundedRectangleBorder(
-            borderRadius: borderRadius,
-            side: borderSide,
-          )
-        : BeveledRectangleBorder(
-            borderRadius: borderRadius,
-            side: borderSide,
-          );
-    var decoration = ShapeDecoration(
-        color: gradientEnable ? null : widget.color,
-        gradient: gradientEnable ? widget.gradient : null,
-        shadows: widget.shadowColor != null && widget.shadowBlur != 0
-            ? [
-                BoxShadow(
-                  color: widget.shadowColor ??
-                      widget.color.withOpacity(disableOpacity),
-                  offset: widget.shadowOffset ?? Offset(0, 0),
-                  blurRadius: widget.shadowBlur,
-                )
-              ]
-            : null,
-        shape: shape);
-    var layerText = _buildTextLayer(_textColor, themeData);
-    var layerRow = _buildRowLayer(layerText);
-    var layerContainer = _buildContainerLayer(layerRow);
-    var layerEffect =
-        _buildEffectLayer(_textColor, shape, borderRadius, layerContainer);
-    var layerDecoratedBox = _buildDecoratedBoxLayer(decoration, layerEffect);
-    var layerGestureDetector = widget.effect
-        ? layerDecoratedBox
-        : GestureDetector(
-            onTap: widget.enabled ? _onPressed : null,
-            child: layerDecoratedBox,
-          );
-    var result = Semantics(
+    FShape shape = FShape(
+      borderShape: widget.cornerStyle == FCornerStyle.round
+          ? FBorderShape.RoundedRectangle
+          : FBorderShape.BeveledRectangle,
+      side: borderSide,
+      borderRadius: borderRadius,
+    );
+
+    /// Handle the relationship between loading text and regular text
+    String loadingText =
+        widget.loadingText == null ? widget.text : widget.loadingText;
+    String text = loading ? loadingText : widget.text;
+    Widget layerText = _buildTextLayer(text);
+    Widget layerRow = _buildRowLayer(text, image, layerText);
+
+    /////////////////
+    Widget layerContainer = FControl(
+      lightOrientation: widget.lightOrientation,
+      width: widget.width,
+      height: widget.height,
+      padding: widget.padding,
+      gradient: widget.gradient,
+      shape: shape,
+      surface: widget.surfaceStyle,
+      supportDropShadow: (shadowColor != null && shadowBlur != 0.0) ||
+          widget.isSupportNeumorphism,
+      dropShadow: FShadow(
+        highlightColor: widget.isSupportNeumorphism
+            ? widget.highlightShadowColor ?? Colors.white.withOpacity(0.83)
+            : Colors.transparent,
+        highlightBlur: widget.isSupportNeumorphism ? _shadowBlur : 0.0,
+        highlightDistance: shadowDistance,
+        shadowColor: shadowColor ?? Color(0xffd1d9e6),
+        shadowBlur: _shadowBlur,
+        shadowDistance: shadowDistance,
+        shadowOffset: widget.shadowOffset,
+      ),
+      supportInnerShadow: widget.isSupportNeumorphism,
+      innerShadow: FShadow(
+        highlightColor:
+            widget.highlightShadowColor ?? Colors.white.withOpacity(0.83),
+        highlightBlur: _shadowBlur,
+        highlightDistance: shadowDistance,
+        shadowColor: shadowColor ?? Color(0xffd1d9e6),
+        shadowBlur: _shadowBlur,
+        shadowDistance: shadowDistance,
+      ),
+      appearance: widget.isSupportNeumorphism
+          ? FAppearance.Neumorphism
+          : FAppearance.Material,
+      onTapCallback: (_, __) {
+        onPressed();
+      },
+      onTapDownCallback: (_, __) {
+        widget.onPressedDown?.call();
+      },
+      onTapUpCallback: (_, __) {
+        widget.onPressedUp?.call();
+      },
+      onTapCancelCallback: (_, __) {
+        widget.onPressedCancel?.call();
+      },
+      maskColor: widget.activeMaskColor,
+      colorForCallback: (sender, state) {
+        if (state == FState.Highlighted) {
+          return widget.isSupportNeumorphism
+              ? widget.color
+              : widget.highlightColor;
+        } else if (state == FState.Disable) {
+          return widget.disabledColor ??
+              (widget.color ?? FDisableColor).withOpacity(disableOpacity);
+        }
+        return widget.color;
+      },
+      disabled: !enabled,
+      hoverColor: widget.hoverColor,
+      onHover: widget.onHover != null
+          ? (v) {
+              widget.onHover(v);
+            }
+          : null,
+      child: Container(
+        alignment: widget.alignment,
+        child: layerRow,
+      ),
+    );
+    /////////////////
+
+    Widget result = Semantics(
       button: true,
-      enabled: widget.enabled,
-      child: layerGestureDetector,
+      enabled: enabled,
+      child: layerContainer,
     );
     return result;
   }
 
-  _buildDecoratedBoxLayer(ShapeDecoration decoration, Widget layerEffect) {
-    var layerDecoratedBox;
-    if (widget.clickEffect) {
-      layerDecoratedBox = AnimatedContainer(
-        duration: Duration(milliseconds: 200),
-        decoration: decoration,
-        onEnd: _onAnimationEnd,
-        child: layerEffect,
-      );
+  double get _shadowBlur {
+    if ((shadowBlur == null || shadowBlur == 0.0) &&
+        widget.isSupportNeumorphism) {
+      return 6.0;
     } else {
-      layerDecoratedBox = DecoratedBox(
-        decoration: decoration,
-        child: layerEffect,
-      );
+      return shadowBlur;
     }
-    return layerDecoratedBox;
   }
 
-  Text _buildTextLayer(Color _textColor, ThemeData themeData) {
+  double get shadowDistance {
+    if (widget.isSupportNeumorphism) {
+      return widget.shadowOffset?.dy ?? 3.0;
+    } else {
+      return widget.shadowOffset?.dy ?? 0.0;
+    }
+  }
+
+  Text _buildTextLayer(String text) {
     return Text(
-      widget.text ?? "",
-      style: TextStyle(
-          color: _textColor,
-          fontWeight: widget.textWeight,
-          height: widget.fontHeight,
-          fontSize: widget.fontSize ?? themeData.textTheme.button.fontSize,
-          fontStyle: widget.fontStyle ?? themeData.textTheme.button.fontStyle),
+      text ?? "",
+      style: enabled ? widget.style : widget.disableStyle ?? widget.style,
     );
   }
 
-  Widget _buildRowLayer(Widget layerText) {
-    if (widget.image == null) return layerText;
-    var showLoading = (widget.loading && widget.hideTextOnLoading);
-    if (showLoading || widget.text == null || widget.text == "") {
-      return widget.image;
+  Widget _buildRowLayer(String text, Widget image, Widget layerText) {
+    if (image == null) return layerText;
+    var showLoading = (loading && widget.hideTextOnLoading);
+    if (showLoading || text == null || text == "") {
+      return image;
     } else {
       switch (widget.imageAlignment) {
         case ImageAlignment.left:
           return Row(
             crossAxisAlignment: CrossAxisAlignment.center,
             mainAxisAlignment: MainAxisAlignment.center,
+            mainAxisSize: MainAxisSize.min,
             children: <Widget>[
-              widget.image,
+              image,
               SizedBox(width: widget.imageMargin),
               layerText,
             ],
@@ -493,8 +522,9 @@ class _FButton extends State<FButton> {
           return Column(
             crossAxisAlignment: CrossAxisAlignment.center,
             mainAxisAlignment: MainAxisAlignment.center,
+            mainAxisSize: MainAxisSize.min,
             children: <Widget>[
-              widget.image,
+              image,
               SizedBox(height: widget.imageMargin),
               layerText,
             ],
@@ -503,20 +533,22 @@ class _FButton extends State<FButton> {
           return Row(
             crossAxisAlignment: CrossAxisAlignment.center,
             mainAxisAlignment: MainAxisAlignment.center,
+            mainAxisSize: MainAxisSize.min,
             children: <Widget>[
               layerText,
               SizedBox(width: widget.imageMargin),
-              widget.image,
+              image,
             ],
           );
         case ImageAlignment.bottom:
           return Column(
             crossAxisAlignment: CrossAxisAlignment.center,
             mainAxisAlignment: MainAxisAlignment.center,
+            mainAxisSize: MainAxisSize.min,
             children: <Widget>[
               layerText,
               SizedBox(height: widget.imageMargin),
-              widget.image,
+              image,
             ],
           );
       }
@@ -524,84 +556,33 @@ class _FButton extends State<FButton> {
     return layerText;
   }
 
-  Container _buildContainerLayer(Widget layerRow) {
-    return Container(
-      width: widget.width,
-      height: widget.height,
-      alignment: Alignment.center,
-      padding: widget.padding ??
-          EdgeInsets.symmetric(horizontal: 16.0, vertical: 16),
-      child: layerRow,
-    );
-  }
-
-  Widget _buildEffectLayer(Color _textColor, ShapeBorder shapeBorder,
-      BorderRadius borderRadius, Container layerContainer) {
-    return Material(
-      clipBehavior: Clip.none,
-      type: MaterialType.transparency,
-      child: widget.effect
-          ? InkWell(
-              onTap: widget.enabled ? _onPressed : null,
-              onHighlightChanged: _handleHighlightChanged,
-              onHover: _handleHoveredChanged,
-              focusNode: widget.focusNode,
-              autofocus: widget.autofocus,
-              splashColor: widget.splashColor ?? _textColor.withOpacity(0.12),
-              highlightColor:
-                  widget.highlightColor ?? _textColor.withOpacity(0.16),
-              hoverColor: widget.hoverColor ?? _textColor.withOpacity(0.04),
-              borderRadius: borderRadius,
-              customBorder: shapeBorder,
-              child: layerContainer,
-            )
-          : layerContainer,
-    );
-  }
-
-  void _onAnimationEnd() {
-    setState(() {
-      widget.shadowBlur = widget.cacheShadowBlur;
-      widget.shadowColor = widget.cacheShadowColor;
-    });
-  }
-
-  _onPressed() {
-    widget.onPressed();
-    if (widget.clickLoading && !widget.loading) {
-      widget.loading = true;
+  onPressed() {
+    widget.onPressed?.call();
+    if (widget.clickLoading && !loading) {
+      loading = true;
       setState(() {});
     } else if (widget.clickEffect) {
-      _pressOutEffect();
-      setState(() {});
+      pressOutEffect();
     }
   }
 
-  _pressOutEffect() {
-    widget.shadowBlur = (widget.shadowBlur ?? 0) + 6;
-    widget.shadowColor = widget.shadowColor ?? widget.color.withOpacity(0.68);
+  pressOutEffect() {
+    shadowBlur = (shadowBlur ?? 0) + 6;
+    shadowColor = widget.shadowColor;
+    if (shadowColor == null) {
+      if (widget.isSupportNeumorphism) {
+        shadowColor = Color(0xffd1d9e6);
+      } else if (widget.strokeColor != null) {
+        shadowColor = widget.strokeColor.withOpacity(0.58);
+      } else {
+        shadowColor = widget.color.withOpacity(0.58);
+      }
+    }
+    shadowTween
+      ..begin = shadowBlur
+      ..end = widget.shadowBlur;
+    animationController
+      ..value = 0.0
+      ..forward();
   }
-}
-
-/// 为 [FButton] 设置圆角。详见 [FButton.corner]
-///
-/// Set rounded corners for [FButton]. See [FButton.corner] for details
-class FButtonCorner {
-  final double leftTopCorner;
-  final double rightTopCorner;
-  final double rightBottomCorner;
-  final double leftBottomCorner;
-
-  const FButtonCorner({
-    this.leftTopCorner = 0,
-    this.rightTopCorner = 0,
-    this.rightBottomCorner = 0,
-    this.leftBottomCorner = 0,
-  });
-
-  FButtonCorner.all(double radius)
-      : leftTopCorner = radius,
-        rightTopCorner = radius,
-        rightBottomCorner = radius,
-        leftBottomCorner = radius;
 }